GalerkinToolkit

Documentation for GalerkinToolkit.

GalerkinToolkit.GenericJaggedArray

GalerkinToolkit.GenericJaggedArray

GalerkinToolkit.JaggedArray

GalerkinToolkit.JaggedArray

GalerkinToolkit.JaggedArray

GalerkinToolkit.jagged_array

GalerkinToolkit.prefix!

GalerkinToolkit.rewind!

GalerkinToolkit.GenericJaggedArray — Type

struct GenericJaggedArray{T,A,B}

Generalization of JaggedArray, where the fields data and ptrs are allowed to be any array-like object.

Properties

data::A
ptrs::B

Supertype hierarchy

GenericJaggedArray{V,A,B} <: AbstractVector{V}

Given a::GenericJaggedArray, V is typeof(view(a.data,a.ptrs[i]:(a.ptrs[i+1]-1))).

GalerkinToolkit.GenericJaggedArray — Method

GenericJaggedArray(data,ptrs)

Create a GenericJaggedArray from the given data and ptrs fields. The resulting object stores references to the given vectors.

GalerkinToolkit.JaggedArray — Type

struct JaggedArray{T,Ti}

Efficient implementation of a vector of vectors. The inner vectors are stored one after the other in consecutive memory locations using an auxiliary vector data. The range of indices corresponding to each inner vector are encoded using a vector of integers ptrs.

Properties

data::Vector{T}
ptrs::Vector{Ti}

Given a::JaggedArray, a.data contains the inner vectors. The i-th inner vector is stored in the range a.ptrs[i]:(a.ptrs[i+1]-1). The number of inner vectors (i.e. length(a)) is length(a.ptrs)-1. a[i] returns a view of a.data restricted to the range a.ptrs[i]:(a.ptrs[i+1]-1).

Supertype hierarchy

JaggedArray{T,Ti} <: AbstractVector{V}

Given a::JaggedArray, V is typeof(view(a.data,a.ptrs[i]:(a.ptrs[i+1]-1))).

GalerkinToolkit.JaggedArray — Method

JaggedArray(a)

Create a JaggedArray object from the vector of vectors a. If a::JaggedArray, then a is returned. Otherwise, the contents of a are copied.

GalerkinToolkit.JaggedArray — Method

JaggedArray(data::Vector,ptrs::Vector)

Create a JaggedArray from the given data and ptrs fields. The resulting object stores references to the given vectors.

GalerkinToolkit.jagged_array — Method

jagged_array(data,ptrs)

Create a JaggedArray or a GenericJaggedArray object depending on the type of data and ptrs. The returned object stores references to the given inputs.

GalerkinToolkit.prefix! — Method

prefix!(a)

Sets a[1]=1 and then a[i+1]=a[i] for i in 1:length(a)

GalerkinToolkit.rewind! — Method

rewind!(a)

Sets a[i+1]=a[i] for i in (length(a)-1):-1:1 and finally a[1]=1.