GalerkinToolkit.AbstractChain — Type

abstract type AbstractChain

Basic queries

node_coordinates

face_nodes

face_reference_id

reference_faces

periodic_nodes

physical_faces

physical_nodes

outwards_normals

Basic constructors

chain_from_arrays

GalerkinToolkit.AbstractFaceGeometry — Type

abstract type AbstractFaceGeometry

Basic queries

num_dims

is_axis_aligned

is_simplex

is_n_cube

real_type

int_type

is_unit_n_cube

is_unit_simplex

is_unitary

bounding_box

boundary

vertex_permutations

Basic constructors

unit_simplex

unit_n_cube

Supertype hierarchy

AbstractFaceGeometry <: GT.AbstractType

GalerkinToolkit.AbstractFaceTopology — Type

abstract type AbstractFaceTopology

Basic queries

boundary

vertex_permutations

Basic constructors

topology

GalerkinToolkit.AbstractMesh — Type

abstract type AbstractMesh

Basic queries

node_coordinates

face_nodes

face_reference_id

reference_faces

periodic_nodes

physical_faces

physical_nodes

outwards_normals

Basic constructors

mesh_from_arrays

mesh_from_gmsh

cartesian_mesh

mesh_from_chain

GalerkinToolkit.AbstractMeshFace — Type

abstract type AbstractMeshFace

Basic queries

geometry

num_dims

num_nodes

node_coordinates

monomial_exponents

primal_basis

dual_basis

lib_to_user_nodes

shape_functions

tabulator

boundary

interior_nodes

interior_node_permutations

Basic constructors

lagrange_mesh_face

GalerkinToolkit.AbstractMeshTopology — Type

abstract type AbstractMeshTopology

Basic queries

face_incidence

face_reference_id

face_permutation_ids

reference_faces

Basic constructors

topology

GalerkinToolkit.boundary — Method

GalerkinToolkit.bounding_box — Method

GalerkinToolkit.cartesian_mesh — Method

GalerkinToolkit.chain_from_arrays — Method

GalerkinToolkit.classify_mesh_nodes! — Function

GalerkinToolkit.complexify — Method

GalerkinToolkit.coordinates — Method

GalerkinToolkit.dual_basis — Method

GalerkinToolkit.face_incidence — Method

GalerkinToolkit.face_nodes — Method

GalerkinToolkit.face_permutation_ids — Method

GalerkinToolkit.face_reference_id — Method

GalerkinToolkit.geometry — Method

GalerkinToolkit.int_type — Method

GalerkinToolkit.interior_node_permutations — Method

GalerkinToolkit.interior_nodes — Method

GalerkinToolkit.is_axis_aligned — Method

GalerkinToolkit.is_n_cube — Method

GalerkinToolkit.is_simplex — Method

GalerkinToolkit.is_unit_n_cube — Method

GalerkinToolkit.is_unit_simplex — Method

GalerkinToolkit.is_unitary — Method

GalerkinToolkit.label_boundary_faces! — Method

label_boundary_faces!(mesh::PMesh;physical_name="boundary")

Update mesh inplace by using partition ownership of faces to label only the boundary of meshes in a parallel mesh where the boundary is defined as a face owned by a given partition and not incident with any other faces on another partition (i.e., not on the interface).

GalerkinToolkit.lagrange_mesh_face — Method

GalerkinToolkit.lib_to_user_nodes — Method

GalerkinToolkit.mesh_from_arrays — Method

GalerkinToolkit.mesh_from_chain — Method

GalerkinToolkit.mesh_from_gmsh — Method

GalerkinToolkit.monomial_exponents — Method

GalerkinToolkit.node_coordinates — Method

GalerkinToolkit.node_permutations — Method

GalerkinToolkit.num_dims — Method

num_dims(a)

Return the number of space dimensions of a. Defaults to a.num_dims.

GalerkinToolkit.num_nodes — Method

GalerkinToolkit.outwards_normals — Method

GalerkinToolkit.periodic_nodes — Method

GalerkinToolkit.physical_faces — Method

GalerkinToolkit.physical_names — Method

GalerkinToolkit.physical_nodes — Method

GalerkinToolkit.primal_basis — Method

GalerkinToolkit.real_type — Method

GalerkinToolkit.reference_faces — Method

GalerkinToolkit.shape_functions — Method

GalerkinToolkit.simplexify — Method

GalerkinToolkit.tabulator — Method

GalerkinToolkit.topology — Method

GalerkinToolkit.unit_n_cube — Method

GalerkinToolkit.unit_simplex — Method

GalerkinToolkit.vertex_permutations — Method

GalerkinToolkit.weights — Method