GalerkinToolkit

Documentation for GalerkinToolkit.

GalerkinToolkit.INVALID_ID

GalerkinToolkit.GenericJaggedArray

GalerkinToolkit.GenericJaggedArray

GalerkinToolkit.JaggedArray

GalerkinToolkit.JaggedArray

GalerkinToolkit.JaggedArray

GalerkinToolkit.ambient_dim

GalerkinToolkit.domain_dim

GalerkinToolkit.face_faces

GalerkinToolkit.face_nodes

GalerkinToolkit.face_own_nodes

GalerkinToolkit.face_ref_id

GalerkinToolkit.face_vertices

GalerkinToolkit.hanging_nodes

GalerkinToolkit.has_hanging_nodes

GalerkinToolkit.has_periodic_nodes

GalerkinToolkit.has_physical_group

GalerkinToolkit.is_hypercube

GalerkinToolkit.is_simplex

GalerkinToolkit.jagged_array

GalerkinToolkit.new_physical_group_id

GalerkinToolkit.node_coordinates

GalerkinToolkit.num_faces

GalerkinToolkit.num_hanging_nodes

GalerkinToolkit.num_nodes

GalerkinToolkit.num_periodic_nodes

GalerkinToolkit.periodic_nodes

GalerkinToolkit.physical_group!

GalerkinToolkit.physical_group_faces

GalerkinToolkit.physical_group_faces!

GalerkinToolkit.physical_group_id

GalerkinToolkit.physical_group_ids

GalerkinToolkit.physical_group_name

GalerkinToolkit.physical_group_names

GalerkinToolkit.physical_groups

GalerkinToolkit.prefix!

GalerkinToolkit.ref_face_faces

GalerkinToolkit.ref_face_nodes

GalerkinToolkit.ref_face_own_nodes

GalerkinToolkit.ref_faces

GalerkinToolkit.rewind!

GalerkinToolkit.INVALID_ID — Constant

const INVALID_ID::Int

Stores an invalid linear index for 1-based indexed objects. INVALID_ID is some number less than 1. The user should not rely on the specific value.

GalerkinToolkit.GenericJaggedArray — Type

struct GenericJaggedArray{V,A,B}

Generalization of JaggedArray, where the fields data and ptrs are allowed to be any array-like object.

Properties

data::A
ptrs::B

Supertype hierarchy

GenericJaggedArray{V,A,B} <: AbstractVector{V}

Given a::GenericJaggedArray, V is typeof(view(a.data,a.ptrs[i]:(a.ptrs[i+1]-1))).

GalerkinToolkit.GenericJaggedArray — Method

GenericJaggedArray(data,ptrs)

Create a GenericJaggedArray from the given data and ptrs fields. The resulting object stores references to the given vectors.

GalerkinToolkit.JaggedArray — Type

struct JaggedArray{T,Ti}

Efficient implementation of a vector of vectors. The inner vectors are stored one after the other in consecutive memory locations using an auxiliary vector data. The range of indices corresponding to each inner vector are encoded using a vector of integers ptrs.

Properties

data::Vector{T}
ptrs::Vector{Ti}

Given a::JaggedArray, a.data contains the inner vectors. The i-th inner vector is stored in the range a.ptrs[i]:(a.ptrs[i+1]-1). The number of inner vectors (i.e. length(a)) is length(a.ptrs)-1. a[i] returns a view of a.data restricted to the range a.ptrs[i]:(a.ptrs[i+1]-1).

Supertype hierarchy

JaggedArray{T,Ti} <: AbstractVector{V}

Given a::JaggedArray, V is typeof(view(a.data,a.ptrs[i]:(a.ptrs[i+1]-1))).

GalerkinToolkit.JaggedArray — Method

JaggedArray(a)

Create a JaggedArray object from the vector of vectors a. If a::JaggedArray, then a is returned. Otherwise, the contents of a are copied.

GalerkinToolkit.JaggedArray — Method

JaggedArray(data::Vector,ptrs::Vector)

Create a JaggedArray from the given data and ptrs fields. The resulting object stores references to the given vectors.

GalerkinToolkit.ambient_dim — Function

ambient_dim(geo)

Return the number of dimensions of the ambient space, where the geometrical object geo is embedded.

GalerkinToolkit.domain_dim — Function

domain_dim(geo)

Return the number of space dimensions of the geometrical object geo.

GalerkinToolkit.face_faces — Function

face_faces(geo,dim_from,dim_to)

Return the face-to-face connectivity of faces of dimension dim_from against faces of dimension dim_to in geo. This is represented with a vector of vectors of outer length num_faces(geo,dim_from). For dim_from > dim_to, the inner vectors contain the face ids of faces of dimension dim_to on the boundary of the corresponding face of dimension dim_from. For dim_from > dim_to, the results is for the co-boundary instead of the boundary. For dim_from == dim_to, the result is the identity.

GalerkinToolkit.face_nodes — Function

face_nodes(geo,dim)

Return the face node connectivity for faces of dimension dim in geo. The result is a vector of vectors of outer length num_faces(geo,dim) that can be indexed with a face id, returning the node ids of the nodes on the closure of this face.

GalerkinToolkit.face_own_nodes — Function

face_own_nodes(geo,dim)

Like face_nodes(geo,dim), but it returns node ids in the interior of the faces in geo.

GalerkinToolkit.face_ref_id — Function

face_ref_id(geo,dim)

Return a vector of integer ids, indicating which is the reference face for each face of dimension dim in the geometrical object geo. I.e.,

id_to_ref_face = ref_faces(geo,dim)
face_to_id = face_ref_id(geo,dim)
ref_face = id_to_ref_face[face_to_id[face]]

ref_face is the reference face associated with face number face of dimension dim in geo.

GalerkinToolkit.face_vertices — Function

face_vertices(geo,dim)

Return the face vertex connectivity for faces of dimension dim in geo. The result is a vector of vectors of outer length num_faces(geo,dim) that can be indexed with a face id, returning the node ids of the nodes on the closure of this face.

GalerkinToolkit.hanging_nodes — Function

info = hanging_nodes(geo)

Returns information about hanging nodes in geo.

info.hanging is a vector with the node ids of hanging nodes.

info.masters is a vector containing the corresponding master ids.

info.coeffs is a vector containing the coefficients to compute

the value at a hanging node from its corresponding masters.

These vectors are of length num_hanging_nodes(geo).

GalerkinToolkit.has_hanging_nodes — Function

has_hanging_nodes(geo)

true if num_hanging_nodes(geo)>0, false otherwise.

GalerkinToolkit.has_periodic_nodes — Function

has_periodic_nodes(geo)

true if num_periodic_nodes(geo)>0, false otherwise.

GalerkinToolkit.has_physical_group — Function

has_physical_group(groups,dim,label)

true if there is a group of dimension dim with label label in groups.

GalerkinToolkit.is_hypercube — Function

is_hypercube(geo)

true if geo represents a hypercube, false otherwise.

GalerkinToolkit.is_simplex — Function

is_simplex(geo)

true if geo represents a simplex, false otherwise.

GalerkinToolkit.jagged_array — Method

jagged_array(data,ptrs)

Create a JaggedArray or a GenericJaggedArray object depending on the type of data and ptrs. The returned object stores references to the given inputs.

GalerkinToolkit.new_physical_group_id — Function

new_physical_group_id(groups)

Find a new available id in groups.

GalerkinToolkit.node_coordinates — Function

node_coordinates(geo)

Return a vector of length num_nodes(geo) with the nodal coordinates of the nodes in geo.

GalerkinToolkit.num_faces — Function

num_faces(geo,dim)

Number of faces of dimension dim in geo.

GalerkinToolkit.num_hanging_nodes — Function

num_hanging_nodes(geo)

Number of hanging nodes in geo.

GalerkinToolkit.num_nodes — Function

num_nodes(geo)

Number of nodes in geo.

GalerkinToolkit.num_periodic_nodes — Function

num_periodic_nodes(geo)

Number of periodic nodes in geo.

GalerkinToolkit.periodic_nodes — Function

info = periodic_nodes(geo)

Returns information about periodic nodes in geo.

info.periodic is a vector with the node ids of periodic nodes.

info.master is a vector containing the corresponding master ids.

info.coeff is a vector containing the scaling a master node over

its periodic node (usually a vector of 1.0).

GalerkinToolkit.physical_group! — Function

physical_group!(groups,dim,name[,id])

Add a new empty physical group to groups of dimension dim name name and id id. If id is not provided, an id will be chosen automatically using new_physical_group_id(groups).

GalerkinToolkit.physical_group_faces — Function

physical_group_faces(groups,dim,label)

Return the face ids for faces of dimension dim with label label in the physical groups groups. The label can be either an integer group id or a string with the group name.

GalerkinToolkit.physical_group_faces! — Function

physical_group_faces!(groups,faces,dim,label)

Set the face ids faces for faces of dimension dim with label label in the physical groups groups. The label can be either an integer group id or a string with the group name.

GalerkinToolkit.physical_group_id — Function

physical_group_id(groups,dim,label)

Return the id of the physical group of dimension dim and label label. The label can be either an integer group id or a string with the group name. In the former case, this function will raise an exception if the given group id is not found.

GalerkinToolkit.physical_group_ids — Function

physical_group_ids(groups,dim)

Return a vector with the ids of physical groups in groups of dimension dim.

GalerkinToolkit.physical_group_name — Function

physical_group_name(groups,dim,label)

Return the name of the physical group of dimension dim and label label. The label can be either an integer group id or a string with the group name. In the latter case, this function will raise an exception if the given group name is not found.

GalerkinToolkit.physical_group_names — Function

physical_group_names(groups,dim)

Return a vector with the names of physical groups in groups of dimension dim.

GalerkinToolkit.physical_groups — Function

groups = physical_groups(geo)

Get a handle to the physical groups in geo.

GalerkinToolkit.prefix! — Method

prefix!(a)

Sets a[1]=1 and then a[i+1]=a[i] for i in 1:length(a).

GalerkinToolkit.ref_face_faces — Method

ref_face_faces(geo,dim,m,n)

Collect the m-face to n-face connectivity for all reference faces of dimension dim in geo.

GalerkinToolkit.ref_face_nodes — Method

ref_face_nodes(geo,dim,n)

Collect the n-face node connectivity for all reference faces of dimension dim in geo.

GalerkinToolkit.ref_face_own_nodes — Method

ref_face_own_nodes(geo,dim,n)

Collect the n-face own node connectivity for all reference faces of dimension dim in geo.

GalerkinToolkit.ref_faces — Function

ref_faces(geo,dim)

Return a vector with the reference faces of dimension dim associated with the geometrical object geo.

GalerkinToolkit.rewind! — Method

rewind!(a)

Sets a[i+1]=a[i] for i in (length(a)-1):-1:1 and finally a[1]=1.