API

GalerkinToolkit.AbstractDomain — Type

abstract type AbstractDomain <: AbstractType end

Abstract type representing the geometry of a single mesh face, typically one of the reference faces.

Basic queries

num_dims

is_axis_aligned

is_simplex

is_n_cube

is_unit_n_cube

is_unit_simplex

is_unitary

bounding_box

vertex_permutations

mesh

faces

inverse_faces

geometries

topology

geometries

options

is_boundary

face_around

Basic constructors

unit_simplex

unit_n_cube

domain

GalerkinToolkit.AbstractMesh — Type

abstract type AbstractMesh

Basic queries

node_coordinates

face_nodes

face_reference_id

reference_spaces

periodic_nodes

physical_faces

outward_normals

Basic constructors

mesh

mesh_from_gmsh

cartesian_mesh

GalerkinToolkit.AbstractQuadrature — Type

abstract type AbstractQuadrature

Basic queries

domain

coordinates

weights

[num_points]

face_reference_id

reference_quadratures

Basic constructors

quadrature

duffy_quadrature

tensor_product_quadrature

node_quadrature

Supertype hierarchy

AbstractQuadrature <: GT.AbstractType

GalerkinToolkit.AbstractSpace — Type

abstract type AbstractSpace <: AbstractType end

Basic queries

domain num_dofs num_nodes face_dofs face_nodes face_reference_id reference_spaces interior_nodes interior_nodes_permutations geometry_own_dofs geometry_own_dofs_permutations geometry_interior_nodes geometry_interior_nodes_permutations geometry_nodes geometry_nodes_permutations

Basic constructors

lagrange_space raviart_thomas_space

GalerkinToolkit.AbstractTopology — Type

abstract type AbstractTopology

Basic queries

face_incidence

face_reference_id

face_permutation_ids

reference_topologies

vertex_permutations

Basic constructors

topology

GalerkinToolkit.AbstractType — Type

abstract type AbstractType end

Parent of all types defined in GalerkinToolkit.

GalerkinToolkit.Options — Type

struct Options{...} <: AbstractType

Type of the objects returned by function options. All properties and type parameters are private.

Basic queries

reference_int_type

int_type

global_int_type

real_type

GalerkinToolkit.bounding_box — Function

GalerkinToolkit.cartesian_mesh — Method

GalerkinToolkit.coordinates — Function

GalerkinToolkit.domain — Function

GalerkinToolkit.duffy_quadrature — Method

GalerkinToolkit.face_around — Function

GalerkinToolkit.face_dofs — Function

GalerkinToolkit.face_incidence — Function

GalerkinToolkit.face_nodes — Function

GalerkinToolkit.face_permutation_ids — Function

GalerkinToolkit.face_reference_id — Function

GalerkinToolkit.faces — Function

GalerkinToolkit.geometries — Function

GalerkinToolkit.geometry_interior_nodes — Function

GalerkinToolkit.geometry_interior_nodes_permutations — Function

GalerkinToolkit.geometry_nodes — Function

GalerkinToolkit.geometry_nodes_permutations — Function

GalerkinToolkit.geometry_own_dofs — Function

GalerkinToolkit.geometry_own_dofs_permutations — Function

GalerkinToolkit.global_int_type — Method

global_int_type(options::Options)

Return the type of the integers used to enumerate global quantities.

GalerkinToolkit.int_type — Method

int_type(options::Options)

Return the default integer type used in the computation except for reference and global quantities.

GalerkinToolkit.interior_node_permutations — Method

GalerkinToolkit.interior_nodes — Function

GalerkinToolkit.interior_nodes_permutations — Function

GalerkinToolkit.inverse_faces — Function

GalerkinToolkit.is_axis_aligned — Function

GalerkinToolkit.is_boundary — Function

GalerkinToolkit.is_n_cube — Function

GalerkinToolkit.is_simplex — Function

GalerkinToolkit.is_unit_n_cube — Function

GalerkinToolkit.is_unit_simplex — Method

GalerkinToolkit.is_unitary — Function

GalerkinToolkit.label_boundary_faces! — Method

label_boundary_faces!(mesh::PMesh;physical_name="boundary")

Update mesh inplace by using partition ownership of faces to label only the boundary of meshes in a parallel mesh where the boundary is defined as a face owned by a given partition and not incident with any other faces on another partition (i.e., not on the interface).

GalerkinToolkit.lagrange_space — Function

GalerkinToolkit.mesh — Method

GalerkinToolkit.mesh_from_gmsh — Method

GalerkinToolkit.node_coordinates — Function

GalerkinToolkit.node_permutations — Method

GalerkinToolkit.node_quadrature — Function

GalerkinToolkit.num_dims — Function

GalerkinToolkit.num_dofs — Function

GalerkinToolkit.num_nodes — Function

GalerkinToolkit.options — Method

options(;kwargs...) -> Options

Create an object representing the default options for the current simulation. This object can be used as an optional argument in several object constructors in GalerkinToolkit, such as the mesh constructors cartesian_mesh and mesh_from_gmsh. In this case, the computations using the generated mesh, will use the given options by default.

GalerkinToolkit.outward_normals — Function

GalerkinToolkit.periodic_nodes — Function

GalerkinToolkit.physical_faces — Function

GalerkinToolkit.physical_names — Method

GalerkinToolkit.push — Function

push(a,ai)

Like push!, but creates a new object to store the result. This function is used to push to immutable collections such as tuples.

GalerkinToolkit.quadrature — Method

GalerkinToolkit.raviart_thomas_space — Function

GalerkinToolkit.real_type — Method

real_type(options::Options)

Return the default real type used in the computation.

GalerkinToolkit.reference_int_type — Method

reference_int_type(options::Options)

Return the type of the integers used to enumerate reference quantities.

GalerkinToolkit.reference_quadratures — Function

GalerkinToolkit.reference_spaces — Function

GalerkinToolkit.reference_topologies — Function

GalerkinToolkit.shape_functions — Method

GalerkinToolkit.tabulator — Method

GalerkinToolkit.tensor_product_quadrature — Method

GalerkinToolkit.topology — Method

GalerkinToolkit.unit_n_cube — Method

unit_n_cube(d)
unit_n_cube(Val(d))

Return an object representing a unit d-cube.

GalerkinToolkit.unit_simplex — Method

unit_simplex(d)
unit_simplex(Val(d))

Return an object representing a unit simplex of dimension d.

GalerkinToolkit.val_parameter — Method

val_parameter(a)

For a::Val{A} it returns A. Otherwise, it returns a.

GalerkinToolkit.vertex_permutations — Function

GalerkinToolkit.weights — Function