API

GalerkinToolkit.AbstractDomain — Type

abstract type AbstractDomain{A} <: AbstractType end

Abstract type representing a subset of $\mathbb{R}^d$, typically $d\in\{0,1,2,3\}$. Domains are defined using an underlying computational mesh. A is an (abstract) type indicating on which type of mesh this domain is defined.

See also AbstractMesh.

Level

Beginner

Basic constructors

unit_simplex

unit_n_cube

domain

interior

boundary

skeleton

Basic queries

num_dims

num_ambient_dims

num_codims

is_axis_aligned

is_simplex

is_n_cube

is_unit_n_cube

is_unit_simplex

is_unitary

bounding_box

vertex_permutations

mesh

faces

inverse_faces

options

is_boundary

face_around

GalerkinToolkit.AbstractFaceDomain — Type

abstract type AbstractFaceDomain <: AbstractDomain{AbstractMesh} end

A domain defined on a single mesh face. Typically used as helper to identify cases that only make sense for a single mesh face.

Level

Advanced

Basic constructors

unit_simplex

unit_n_cube

GalerkinToolkit.AbstractFaceSpace — Type

abstract type AbstractFaceSpace <: AbstractSpace{AbstractMesh} end

Like AbstractSpace, but for a single mesh face. Typically used as helper to identify cases that only make sense for a single mesh face.

Level

Advanced

GalerkinToolkit.AbstractFaceTopology — Type

abstract type AbstractFaceTopology <: AbstractTopology end

Like AbstractTopology, but for a single mesh face. Typically used as helper to identify cases that only make sense for a single mesh face.

Level

Advanced

GalerkinToolkit.AbstractMesh — Type

abstract type AbstractMesh <: AbstractType end

Abstract type representing a triangulation of a subset of $\mathbb{R}^d$, typically $d\in\{0,1,2,3\}$, plus metadata useful in finite element computations, such as physical groups for imposing boundary conditions.

Notation

Each of of the elements of the triangulation is referred to as a face. A mesh can contain faces of different dimensions. A mesh might or might not represent a cell complex (all possible low dimensional faces are present in the mesh), but it is often assumed that it represents a cell complex.

Level

Beginner

Basic constructors

mesh

chain

mesh_from_msh

mesh_from_space

cartesian_mesh

complexify

simplexify

Basic queries

num_dims

num_ambient_dims

num_codims

num_faces

num_nodes

node_coordinates

face_nodes

face_reference_id

reference_spaces

periodic_nodes

physical_faces

geometries

outward_normals

GalerkinToolkit.AbstractPMesh — Type

GalerkinToolkit.AbstractQuadrature — Type

abstract type AbstractQuadrature

Basic queries

domain

coordinates

weights

num_points

face_reference_id

reference_quadratures

Basic constructors

quadrature

duffy_quadrature

tensor_product_quadrature

node_quadrature

Supertype hierarchy

AbstractQuadrature <: GT.AbstractType

GalerkinToolkit.AbstractSpace — Type

abstract type AbstractSpace{A} <: AbstractType end

Abstract type representing a finite element space. A is an (abstract) type indicating on which type of mesh this domain is defined.

Level

Basic

Basic constructors

lagrange_space raviart_thomas_space

Basic queries

domain

num_dofs

face_dofs

face_nodes

face_reference_id

reference_spaces

geometry_own_dofs

geometry_own_dofs_permutations

Additional queries

For spaces, used as reference spaces in AbstractMesh specializations.

num_nodes

interior_nodes

interior_nodes_permutations

geometry_interior_nodes

geometry_interior_nodes_permutations

geometry_nodes

geometry_nodes_permutations

GalerkinToolkit.AbstractTopology — Type

abstract type AbstractTopology

Abstract type representing the incidence relations in a cell complex.

See also AbstractFaceTopology.

Level

Intermediate

Basic constructors

topology

Basic queries

face_incidence

face_reference_id

face_permutation_ids

reference_topologies

vertex_permutations

GalerkinToolkit.AbstractType — Type

abstract type AbstractType end

Parent of all types defined in GalerkinToolkit.

GalerkinToolkit.Options — Type

struct Options{...} <: AbstractType

Type of the objects returned by function options. All properties and type parameters are private.

Basic queries

reference_int_type

int_type

global_int_type

real_type

GalerkinToolkit.boundary — Function

GalerkinToolkit.bounding_box — Function

p0,p1 = bounding_box(x)

Return a tuple of two vectors, where the vectors p0 and p1 define the span of the bounding box of x.

Level

Intermediate

GalerkinToolkit.cartesian_mesh — Method

GalerkinToolkit.chain — Method

GalerkinToolkit.complexify — Function

complexify(x)

Convert x into a mesh representing a cell complex.

Level

Intermediate

GalerkinToolkit.coordinates — Function

GalerkinToolkit.domain — Function

GalerkinToolkit.duffy_quadrature — Method

GalerkinToolkit.face_around — Function

face_around(x)

Return an integer that allows to break ties when faces in x need to point to faces around of one dimension higher. Return nothing if x does not break such ties.

Note: This function will eventually return a vector of integers.

Level

Intermediate

GalerkinToolkit.face_dofs — Function

GalerkinToolkit.face_incidence — Function

face_incidence(x,d)

GalerkinToolkit.face_nodes — Function

GalerkinToolkit.face_permutation_ids — Function

GalerkinToolkit.face_reference_id — Function

GalerkinToolkit.faces — Function

faces(x)

Return the subset of face ids in mesh(x) of dimension num_dims(x) defining the domain x. This is effectively the map from domain face id to mesh face id.

See also inverse_faces.

Level

Intermediate

GalerkinToolkit.geometries — Function

geometries(x,d)
geometries(x,Val(d))

Return a vector of domains representing the geometrical entities of x of dimension d. The returned domains and x are defined on the same mesh. That is, faces(geometries(x,1)[2]) are the face ids in mesh(x) representing the second edge of x.

Notation

geometries(x,Val(0)) are referred to as the vertices of x. geometries(x,Val(1)) are referred to as the edges of x. geometries(x,Val(d)) are referred to as the d-faces of x.

Level

Advanced

GalerkinToolkit.geometry_interior_nodes — Function

GalerkinToolkit.geometry_interior_nodes_permutations — Function

GalerkinToolkit.geometry_nodes — Function

GalerkinToolkit.geometry_nodes_permutations — Function

GalerkinToolkit.geometry_own_dofs — Function

GalerkinToolkit.geometry_own_dofs_permutations — Function

GalerkinToolkit.global_int_type — Method

global_int_type(options::Options)

Return the type of the integers used to enumerate global quantities.

GalerkinToolkit.int_type — Method

int_type(options::Options)

Return the default integer type used in the computation except for reference and global quantities.

GalerkinToolkit.interior — Function

GalerkinToolkit.interior_node_permutations — Method

GalerkinToolkit.interior_nodes — Function

GalerkinToolkit.interior_nodes_permutations — Function

GalerkinToolkit.inverse_faces — Function

inverse_faces(x)

Return the inverse integer mas of faces(x). This is effectively the map from mesh face id to domain face id. Mesh faces not present in the domain, receive an invalid index id.

See also faces.

Level

Intermediate

GalerkinToolkit.is_axis_aligned — Function

is_axis_aligned(x)

True if x is a unit simplex or a unit cube.

Level

Advanced

GalerkinToolkit.is_boundary — Function

is_boundary(x)

True if x represent an (internal) boundary. Faces in an internal boundary "point" to only of the two faces around of a dimension higher.

See also face_around.

Level

Intermediate

GalerkinToolkit.is_n_cube — Function

is_n_cube(x)

True if x is a n-cube (hypercube).

See also is_simplex, is_unit_simplex, is_unit_n_cube.

Level

Intermediate

GalerkinToolkit.is_simplex — Function

is_simplex(x)

True if x is a simplex.

See also is_n_cube, is_unit_simplex, is_unit_n_cube.

Level

Intermediate

GalerkinToolkit.is_unit_n_cube — Function

is_unit_n_cube(x)

True if x is a unit n-cube.

See also is_n_cube, is_unit_simplex, is_simplex.

Level

Intermediate

GalerkinToolkit.is_unit_simplex — Function

is_unit_simplex(x)

True if x is a unit simplex.

See also is_n_cube, is_unit_n_cube, is_simplex.

Level

Intermediate

GalerkinToolkit.is_unitary — Function

is_unitary(x)

True bounding_box(x) coincides with a unit n-cube.

Level

Advanced

GalerkinToolkit.lagrange_space — Function

GalerkinToolkit.mesh — Method

GalerkinToolkit.mesh_from_msh — Method

GalerkinToolkit.mesh_from_space — Function

mesh_from_space(space)

Return the mesh induced by space. For instance, a (high order) Lagrange space can be interpreted as a mesh using this function.

Level

Advanced

GalerkinToolkit.node_coordinates — Function

node_coordinates(x)

Return the vector of node coordinates associated with `x.

GalerkinToolkit.node_permutations — Method

GalerkinToolkit.node_quadrature — Function

GalerkinToolkit.num_ambient_dims — Function

num_ambient_dims(x)

Return the ambient dimension where object x lives.

See also num_codims, num_dims.

Level

Beginner

GalerkinToolkit.num_codims — Function

num_codims(x)

Return num_ambient_dims(x)-num_dims(x).

See also num_ambient_dims, num_dims.

Level

Beginner

GalerkinToolkit.num_dims — Function

num_dims(x)

Return the parametric dimension of x.

See also num_ambient_dims, num_codims.

Level

Beginner

GalerkinToolkit.num_dofs — Function

num_dofs(x)

Return the number of degrees of freedom of x.

See also num_nodes.

Level

Beginner

GalerkinToolkit.num_faces — Function

num_faces(x)
num_faces(x,d)

Return the number of faces of dimension d in mesh(x). If d is omitted, return a vector with the number of faces in each dimension, starting from dimension 0 up to num_dims(x).

Level

Beginner

GalerkinToolkit.num_nodes — Function

num_nodes(x)

Return the number of nodes of x.

See also num_dofs.

Level

Beginner

GalerkinToolkit.num_points — Function

num_points(x)

Return the number of integration points in x.

See also num_nodes.

Level

Beginner

GalerkinToolkit.options — Method

options(;kwargs...) -> Options

Create an object representing the default options for the current simulation. This object can be used as an optional argument in several object constructors in GalerkinToolkit, such as the mesh constructors cartesian_mesh and mesh_from_msh. In this case, the computations using the generated mesh, will use the given options by default.

GalerkinToolkit.outward_normals — Function

GalerkinToolkit.periodic_nodes — Function

GalerkinToolkit.physical_faces — Function

GalerkinToolkit.physical_names — Method

GalerkinToolkit.push — Function

push(a,ai)

Like push!, but creates a new object to store the result. This function is used to push to immutable collections such as tuples.

GalerkinToolkit.quadrature — Method

GalerkinToolkit.raviart_thomas_space — Function

GalerkinToolkit.real_type — Method

real_type(options::Options)

Return the default real type used in the computation.

GalerkinToolkit.reference_int_type — Method

reference_int_type(options::Options)

Return the type of the integers used to enumerate reference quantities.

GalerkinToolkit.reference_quadratures — Function

GalerkinToolkit.reference_spaces — Function

GalerkinToolkit.reference_topologies — Function

reference_topologies(x)
reference_topologies(x,d)
reference_topologies(x,Val(d))

Return the list (a vector or a tuple) of reference topologies in x of dimension d. If the second argument is omitted, return a tuple with the reference topologies in each dimension, starting from dimension 0 up to num_dims(x).

See also face_reference_id.

Level

Intermediate

GalerkinToolkit.shape_functions — Method

GalerkinToolkit.simplexify — Function

simplexify(x)

Convert x into a mesh made of simplex cells.

Level

Intermediate

GalerkinToolkit.skeleton — Function

GalerkinToolkit.tabulator — Method

GalerkinToolkit.tensor_product_quadrature — Method

GalerkinToolkit.topology — Method

GalerkinToolkit.unit_n_cube — Method

unit_n_cube(d)
unit_n_cube(Val(d))

Return an object representing a unit d-cube.

GalerkinToolkit.unit_simplex — Method

unit_simplex(d)
unit_simplex(Val(d))

Return an object representing a unit simplex of dimension d.

GalerkinToolkit.val_parameter — Method

val_parameter(a)

For a::Val{A} it returns A. Otherwise, it returns a.

GalerkinToolkit.vertex_permutations — Function

vertex_permutations(x)

Return a list of permutations representing the admissible re-labelings of the vertices of x.

Level

Advanced

GalerkinToolkit.weights — Function